feat: Initialize Cloud9 configuration with Terraform
A basic configuration file for using Cloud9 with Terraform was written.

Basically, created by referring to the official Terraform
`aws_cloud9_environment_ec2` resource at the following URL.

※URL: https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/cloud9_environment_ec2

The version of Terraform to be used is 1.1.7,
which is the latest version available at this time.

※ref: https://github.com/hashicorp/terraform/releases/tag/v1.1.7

As the configuration of Terraform, `main.tf` is used as an entry point,
and the concerns are split into some files.

This is an attempt to follow the `minimal recommended module`
in the official Terraform documentation.

※ref: https://www.terraform.io/language/modules/develop