Introduce CircleCI for checking `terraform fmt` and `terraform validate`
Ensure that the codebase is in a state that `terraform fmt` and
`terraform validate` commands are executed.

To do this, run `terraform fmt` and `terraform validate` commands
on CircleCI and display the results in CI.

When `terraform fmt` failed, tell CI about the failure and make
the error displayed, so use the following command.

```
terraform fmt -diff=true -check=true
```